Moved TopMenu inside Layout
